+++ src/main/java/lotto/domain/winning/WinningStatics.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winning;
 
 import java.util.ArrayList;
 import java.util.EnumMap;
